added optimization settings test and commands
