+++ cypress/support/stackable/commands/index.js
@@ -30,4 +30,5 @@ import './controls'
 import './editor'
 import './global-settings'
 import './inspector'
+import './settings'
